core-master/requirements/prod.txt
amqp==5.1.1
asgiref==3.6.0
billiard==3.6.4.0
celery==5.2.7
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cryptography==38.0.4
defusedxml==0.7.1
diagrams==0.23.1
Django==3.2.8
django-celery-beat==2.4.0
django-filter==22.1
django-timezone-field==5.0
djangorestframework==3.14.0
graphviz==0.19.2
idna==3.4
Jinja2==3.1.2
kombu==5.2.4
MarkupSafe==2.1.1
oauthlib==3.2.2
Pillow==9.3.0
prompt-toolkit==3.0.36
psycopg2-binary==2.9.5
pycparser==2.21
PyJWT==2.6.0
python-crontab==2.7.1
python-dateutil==2.8.2
python-decouple==3.6
python3-openid==3.2.0
pytz==2022.7
requests==2.28.1
requests-oauthlib==1.3.1
sentry-sdk==1.12.1
six==1.16.0
social-auth-app-django==5.0.0
social-auth-core==4.3.0
socialscore==2.1
sqlparse==0.4.3
tweepy==4.12.1
tzdata==2022.7
urllib3==1.26.13
vine==5.0.0
wcwidth==0.2.5
redis==3.5.3